Move reduceToExcerpt() function to common location
Resolves #5
